requirements.txt
boto3
Pillow
tk
opencv-python